refactored general 3D conv kernel tuning functions into their own file for easier reuse across other 3D conv solvers
